add tests for another api models
